[day09] part 2 is a "double recursion" of part 1
instead of simply returning multiply*count, we need to check whether
any of the characters within the count range contain any markers. so,
we call the function twice - first to evaluate what's left in this small
substring to be considered, and then, as before, to evaluate the rest
of so far uncompressed string.