[day11] wip generate next states and hashability
set is not a hashable type, so I can't use sets on State if I define
state using sets. Ugh. This is why I use lists for states and I think
this ultimately slows things down. so I need to go back to the original
implementation where I use lists instead of sets for floors..