[day14] use dicts to keep generated hashes
this speeds things up. instead of generating new hashes every time, we
first generate a list of hashes, then we filter on those which satisfy
the 'has three consec. letters' condition, and then we additionally
filter on those which satisfy 'has five consec. letters'. the latter is
a subset of the former, because if a hash contains five consec. letters,
it definitely contains three - no need to search through the original
list of hashes.

the implementation then loops through the 'threes' dictionary and checks
whether there exists a hash which contains five consec. characters, such
that the repeated character is the same as the one found in the threes.
this is why I keep track of the character itself as well.

part two is super slow but at least it runs, compared to my original
attempt.