[day18] further simplification of safe condition
no need for a function call, we simply need to check for symmetric
conditions, ie. that the first character equals the third character.
The most time is spent on creating a new list with True prepended and
appended to the original list to take into account the 'implied' spots.

I also switched to lists of booleans, that also reduced the run time a
bit.